adding iOS XCFramework support
