--- ACPAnalytics/Assets/Editor.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: a3624a9ecf2f74bceb87ba351cee3ab5
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
