--- ACPAnalytics/Assets/Editor/BuildPostProcessor.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 4cf4ad8b0046846d7abe01b811837337
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
