+++ ACPAnalytics/Assets/Plugins/iOS/Editor.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: a3624a9ecf2f74bceb87ba351cee3ab5
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
