+++ ACPAnalytics/ProjectSettings/ProjectVersion.txt
@@ -1,2 +1,2 @@
-m_EditorVersion: 2019.3.10f1
-m_EditorVersionWithRevision: 2019.3.10f1 (5968d7f82152)
+m_EditorVersion: 2019.4.34f1
+m_EditorVersionWithRevision: 2019.4.34f1 (6a9faed444f2)
